refactor: split command tests from implementation
