+++ crates/core/src/commands/list_apps.rs
@@ -15,41 +15,5 @@ pub fn execute(args: ListAppsArgs, adapter: &dyn PlatformAdapter) -> Result<Valu
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{adapter::PlatformAdapter, error::AdapterError, node::AppInfo};
-
-    struct AppsAdapter;
-
-    impl PlatformAdapter for AppsAdapter {
-        fn list_apps(&self) -> Result<Vec<AppInfo>, AdapterError> {
-            Ok(vec![
-                AppInfo {
-                    name: "Finder".into(),
-                    pid: 1,
-                    bundle_id: Some("com.apple.finder".into()),
-                },
-                AppInfo {
-                    name: "TextEdit".into(),
-                    pid: 2,
-                    bundle_id: Some("com.apple.TextEdit".into()),
-                },
-            ])
-        }
-    }
-
-    #[test]
-    fn app_filter_matches_by_name_case_insensitively() {
-        let value = execute(
-            ListAppsArgs {
-                app: Some("text".into()),
-            },
-            &AppsAdapter,
-        )
-        .unwrap();
-
-        let apps = value["apps"].as_array().unwrap();
-        assert_eq!(apps.len(), 1);
-        assert_eq!(apps[0]["name"], "TextEdit");
-    }
-}
+#[path = "list_apps_tests.rs"]
+mod tests;
