+++ src/command_contract_tests.rs
@@ -3,16 +3,9 @@ use clap::CommandFactory;
 use std::collections::BTreeSet;
 use std::path::{Path, PathBuf};
 
-const NON_COMMAND_MODULES: &[&str] = &[
-    "helpers",
-    "mod",
-    "ref_policy_tests",
-    "resolved_element",
-    "search_text",
-    "wait_tests",
-];
+const NON_COMMAND_MODULES: &[&str] = &["helpers", "mod", "search_text"];
 
-const INLINE_COMMAND_TESTS: &[&str] = &[
+const COMMAND_SPECIFIC_TESTS: &[&str] = &[
     "find",
     "focus-window",
     "is",
@@ -95,6 +88,9 @@ fn every_core_command_module_is_registered_or_declared_helper() {
         .collect();
 
     for module in command_module_files() {
+        if module.ends_with("_tests") {
+            continue;
+        }
         assert!(
             command_modules.contains(module.as_str())
                 || NON_COMMAND_MODULES.contains(&module.as_str()),
@@ -110,6 +106,23 @@ fn every_cli_subcommand_has_explicit_test_coverage_classification() {
         assert!(
             covered.contains(command.as_str()),
             "{command} has no explicit command coverage classification"
+        );
+    }
+}
+
+#[test]
+fn core_command_modules_keep_tests_in_sibling_files() {
+    for path in command_module_paths() {
+        let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
+        if stem.ends_with("_tests") || NON_COMMAND_MODULES.contains(&stem) {
+            continue;
+        }
+
+        let source = std::fs::read_to_string(&path).expect("command module should be readable");
+        assert!(
+            !source.contains("mod tests {"),
+            "{} keeps inline tests; move them to {stem}_tests.rs and include with #[path]",
+            path.display()
         );
     }
 }
@@ -136,19 +149,29 @@ fn command_module_path(module: &str) -> PathBuf {
 }
 
 fn command_module_files() -> Vec<String> {
-    let dir = Path::new(env!("CARGO_MANIFEST_DIR")).join("../crates/core/src/commands");
-    let mut files: Vec<String> = std::fs::read_dir(dir)
-        .expect("command module directory exists")
-        .filter_map(Result::ok)
-        .filter_map(|entry| entry.path().file_stem().map(|stem| stem.to_owned()))
+    let mut files: Vec<String> = command_module_paths()
+        .into_iter()
+        .filter_map(|path| path.file_stem().map(|stem| stem.to_owned()))
         .filter_map(|stem| stem.to_str().map(ToOwned::to_owned))
         .collect();
     files.sort();
     files
 }
 
+fn command_module_paths() -> Vec<PathBuf> {
+    let dir = Path::new(env!("CARGO_MANIFEST_DIR")).join("../crates/core/src/commands");
+    let mut paths: Vec<PathBuf> = std::fs::read_dir(dir)
+        .expect("command module directory exists")
+        .filter_map(Result::ok)
+        .map(|entry| entry.path())
+        .filter(|path| path.extension().is_some_and(|ext| ext == "rs"))
+        .collect();
+    paths.sort();
+    paths
+}
+
 fn coverage_names() -> BTreeSet<&'static str> {
-    INLINE_COMMAND_TESTS
+    COMMAND_SPECIFIC_TESTS
         .iter()
         .chain(SHARED_REF_ACTION_TESTS)
         .chain(BINARY_CONTRACT_TESTS)
